requirements.txt
Flask
numpy
Pillow
scikit-learn
tensorflow
pywebview